+++ src/ascat/plot/__init__.py
@@ -1,0 +1,5 @@
+# SPDX-License-Identifier: MIT
+# SPDX-FileCopyrightText: Copyright (c) 2026 TU Wien
+# SPDX-FileContributor: For a full list of authors, see the AUTHORS file.
+
+from .plot import plot_szf